azure-functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
alembic
bcrypt
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Marshmallow
Flask-Migrate
Flask-SQLAlchemy
marshmallow
marshmallow-sqlalchemy
mysqlclient
PyJWT
PyMySQL
SQLAlchemy
pyodbc
